open_infra/requirements.txt
APScheduler==3.6.3
argcomplete==2.0.0
arrow==1.2.2
attrs==21.4.0
beautifulsoup4==4.11.1
bilibili-api==2.1.4
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssutils==2.3.1
distlib==0.3.4
Django==3.2.14
django-ckeditor==5.8.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-js-asset==1.2.2
django-redis==5.2.0
django-simpleui==3.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docker==2.6.1
docker-compose==1.18.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.17.0
esdk-obs-python==3.22.2
et-xmlfile==1.1.0
filelock==3.4.1
frozenlist==1.2.0
hanlp-downloader
huaweicloudsdkaom
huaweicloudsdkbms
huaweicloudsdkcore
huaweicloudsdkecs
huaweicloudsdkeip
huaweicloudsdkelb
huaweicloudsdknat
huaweicloudsdkrds
huaweicloudsdkvpc
huaweicloudsdkiam
huaweicloudsdkbss
huaweicloudsdkbssintl
huaweicloudsdkswr
icalendar==4.0.9
idna==2.8
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
jenkinsapi==0.3.11
jieba==0.42.1
Jinja2==2.10.3
joblib==1.1.0
JPype1==0.7.0
jsonschema==2.5.1
lxml==4.5.2
Markdown==3.1.1
MarkupSafe==1.1.1
multi-key-dict==2.0.3
multidict==5.2.0
nltk==3.6.7
openpyxl==3.0.10
packaging==19.2
pbr==5.9.0
phone==0.4.3
phonenumbers==8.12.52
Pillow==6.2.1
platformdirs==2.4.0
pycparser==2.19
pycryptodome==3.10.1
pycryptodomex==3.9.4
pyhanlp==0.1.84
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.5
PySocks==1.6.8
python-alipay-sdk==2.0.1
python-dateutil==2.8.2
python-jenkins==1.7.0
python-stdnum==1.17
pytz==2019.3
PyYAML==6.0
redis==3.5.3
regex==2022.7.9
requests==2.24.0
requests-futures==1.0.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
Serializer==0.2.1
simplejson==3.17.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.3.0
textblob==0.17.1
texttable==1.6.2
tqdm==4.64.0
typing_extensions==4.1.1
tzlocal==2.1
uritemplate==3.0.0
urllib3==1.25.7
virtualenv==20.14.1
websocket-client==0.47.0
websockets==9.1
Werkzeug==1.0.1
wget==3.2
XlsxWriter==3.0.3
xmltodict==0.13.0
yarl==1.7.2
zipp==3.6.0
uwsgi